+++ server/lib/lead/post-leads.js
@@ -12,13 +12,12 @@ export default function postLeads(ctx: Object, leads: Array<Object>): Promise {
   client.logger.debug("postLeads", leads.length);
 
   return Promise.map(leads, lead => {
-    return service.intercomClient.post("/contacts")
-      .send(lead)
+    return service.intercomClient.post("/contacts", lead)
       .then(response => response.body)
       .catch(err => {
         const fErr = service.intercomClient.handleError(err);
         client.asUser({ email: lead.email, external_id: lead.user_id }).logger.error("outgoing.user.error", fErr);
         return Promise.resolve(fErr);
       });
-  }, { concurrency: 5 });
+  }, { concurrency: 1 });
 }